upgraded role scoper

+++ plugins/role-scoper/admin/admin_lib_rs.php
@@ -461,8 +461,14 @@ class ScoperAdminLib {
 			
 		$user_clause = "user_id IN ('" . implode("', '", $user_ids) . "')";
 		
-		$table_name = ( $blog_id_arg ) ? $wpdb->base_prefix . $blog_id_arg . '_' . 'user2role2object_rs' : $wpdb->user2role2object_rs;
+		if ( $blog_id_arg )
+			$table_name = ( $blog_id_arg > 1 ) ? $wpdb->base_prefix . $blog_id_arg . '_' . 'user2role2object_rs' : $wpdb->base_prefix . 'user2role2object_rs';
+		else
+			$table_name = $wpdb->user2role2object_rs;
 		
+		if ( ! $wpdb->get_results( "SHOW TABLES LIKE '$table_name'" ) )
+			return;
+
 		scoper_query("DELETE FROM $table_name WHERE $user_clause");
 		
 		foreach ( $user_ids as $user_id ) {
